pyproject.toml
[project]
name = "re-cc"
version = "0.1.0"
description = "Multi-LLM terminal-based coding assistant"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE.md"}
authors = [
    {name = "Re-CC Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.4.2",
    "httpx>=0.24.1",
    "pydantic>=2.0.0",
    "gitpython>=3.1.30",
    "prompt-toolkit>=3.0.38",
    "openai>=1.0.0",
    "anthropic>=0.21.0",
    "ollama>=0.1.0",
    "textual>=0.45.1",
    "keyring>=24.3.0",
    "platformdirs>=4.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.103.1",
    "uvicorn>=0.23.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "mypy>=1.4.1",
    "ruff>=0.0.285",
    "pytest-cov>=4.1.0",
]

[project.scripts]
re-cc = "re_cc.__main__:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "B", "C4", "SIM", "UP"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
